telegrambot/requirements.txt
aiogram>=3.4.1
python-pptx>=0.6.23
python-dotenv>=1.0.1
requests>=2.32.0
pillow>=10.4.0
aiosqlite>=0.20.0
yookassa>=2.4.0
pdf2docx>=0.5.6
